requirements.txt
matplotlib==3.7.2
mysql-connector-python
numpy==1.25.2
pandas==2.0.3
SQLAlchemy==2.0.7
statistics